+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased][unreleased]
 
+* Bugfix: for large braids maxit could overflow an int32 (issue #138).
+
 * Move "utility" debug messages to debug level 2.  Now debug level 1
   should be reserved for mathematical diagnostic information (e.g.,
   internal iterations, convergence).  This provides a cleaner output